Update tests for new behavior (WIP)
