+++ internal/summoner/acquire/utils_test.go
@@ -1,0 +1,13 @@
+package acquire
+
+import (
+    "github.com/spf13/viper"
+    "github.com/stretchr/testify/assert"
+    "net/http"
+    "net/http/httptest"
+    "testing"
+    "time"
+)
+
+func TestGetRobotsTxt(t *testing.T) {
+}
